extract `format_as_text()` from `statement()`
We're now very close to being able to easily add html
formatting.

Now that `statement()` fully separates the collection
of invoice data into an Invoice, and then uses that
Invoice to drive the text formatting of the statement,
extract that formatting code into it's own function.

`format_as_text()` is still called directly from `statement()`,
which allows our tests to continue passing, but in the next
step I'll change `statement()` (and maybe rename it) so it just
returns the Invoice.